# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AuthEngine",
    "description": "Production-ready IAM system built with FastAPI — supports authentication, multi-tenancy, fine-grained permissions, and token introspection to power multiple apps from a unified identity layer.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/platform/tenants/tenants/{tenant_id}": {
      "get": {
        "tags": ["platform-tenants"],
        "summary": "Get Tenant",
        "description": "View tenant details.",
        "operationId": "get_tenant_api_v1_platform_tenants_tenants__tenant_id__get",
        "security": [{ "HTTPBearer": [] }],
        "parameters": [
          {
            "name": "tenant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Tenant Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/TenantResponse" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["platform-tenants"],
        "summary": "Update Tenant",
        "description": "Update tenant details.",
        "operationId": "update_tenant_api_v1_platform_tenants_tenants__tenant_id__put",
        "security": [{ "HTTPBearer": [] }],
        "parameters": [
          {
            "name": "tenant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Tenant Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/TenantUpdate" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/TenantResponse" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["platform-tenants"],
        "summary": "Delete Tenant",
        "description": "Delete tenant.",
        "operationId": "delete_tenant_api_v1_platform_tenants_tenants__tenant_id__delete",
        "security": [{ "HTTPBearer": [] }],
        "parameters": [
          {
            "name": "tenant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Tenant Id"
            }
          }
        ],
        "responses": {
          "204": { "description": "Successful Response" },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/tenants/{tenant_id}/social-providers": {
      "get": {
        "tags": ["tenant-social-providers"],
        "summary": "List tenant social providers",
        "operationId": "list_social_providers_api_v1_tenants__tenant_id__social_providers_get",
        "security": [{ "HTTPBearer": [] }],
        "parameters": [
          {
            "name": "tenant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Tenant Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TenantSocialProviderResponse"
                  },
                  "title": "Response List Social Providers Api V1 Tenants  Tenant Id  Social Providers Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["tenant-social-providers"],
        "summary": "Create tenant social provider credentials",
        "operationId": "create_social_provider_api_v1_tenants__tenant_id__social_providers_post",
        "security": [{ "HTTPBearer": [] }],
        "parameters": [
          {
            "name": "tenant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Tenant Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TenantSocialProviderCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TenantSocialProviderResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/api/v1/tenants/{tenant_id}/social-providers/{provider}": {
      "put": {
        "tags": ["tenant-social-providers"],
        "summary": "Update tenant social provider credentials",
        "operationId": "update_social_provider_api_v1_tenants__tenant_id__social_providers__provider__put",
        "security": [{ "HTTPBearer": [] }],
        "parameters": [
          {
            "name": "tenant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Tenant Id"
            }
          },
          {
            "name": "provider",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Provider" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TenantSocialProviderUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TenantSocialProviderResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["tenant-social-providers"],
        "summary": "Delete tenant social provider credentials",
        "operationId": "delete_social_provider_api_v1_tenants__tenant_id__social_providers__provider__delete",
        "security": [{ "HTTPBearer": [] }],
        "parameters": [
          {
            "name": "tenant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Tenant Id"
            }
          },
          {
            "name": "provider",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Provider" }
          }
        ],
        "responses": {
          "204": { "description": "Successful Response" },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/api/v1/tenants/{tenant_id}/social-providers/{provider}/toggle": {
      "patch": {
        "tags": ["tenant-social-providers"],
        "summary": "Toggle tenant social provider active state",
        "operationId": "toggle_social_provider_api_v1_tenants__tenant_id__social_providers__provider__toggle_patch",
        "security": [{ "HTTPBearer": [] }],
        "parameters": [
          {
            "name": "tenant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Tenant Id"
            }
          },
          {
            "name": "provider",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Provider" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TenantSocialProviderToggle"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TenantSocialProviderResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": { "$ref": "#/components/schemas/ValidationError" },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SocialProviderName": {
        "type": "string",
        "enum": ["google", "microsoft", "github", "authengine"],
        "title": "SocialProviderName"
      },
...
      "TenantResponse": {
        "properties": {
          "name": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Name"
          },
          "description": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Description"
          },
          "type": {
            "anyOf": [
              { "$ref": "#/components/schemas/TenantType" },
              { "type": "null" }
            ]
          },
          "owner_id": {
            "type": "string",
            "format": "uuid",
            "title": "Owner Id"
          },
          "id": { "type": "string", "format": "uuid", "title": "Id" },
          "created_by": {
            "type": "string",
            "format": "uuid",
            "title": "Created By"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "owner_id",
          "id",
          "created_by",
          "created_at",
          "updated_at"
        ],
        "title": "TenantResponse"
      },
...
      "TenantSocialProviderCreate": {
        "properties": {
          "provider": { "$ref": "#/components/schemas/SocialProviderName" },
          "client_id": {
            "type": "string",
            "minLength": 1,
            "title": "Client Id"
          },
          "client_secret": {
            "type": "string",
            "minLength": 1,
            "title": "Client Secret"
          },
          "redirect_uri": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Redirect Uri"
          },
          "oidc_discovery_url": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Oidc Discovery Url"
          }
        },
        "type": "object",
        "required": ["provider", "client_id", "client_secret"],
        "title": "TenantSocialProviderCreate",
        "description": "Body for POST /tenants/{tenant_id}/social-providers"
      },
      "TenantSocialProviderResponse": {
        "properties": {
          "id": { "type": "string", "format": "uuid", "title": "Id" },
          "tenant_id": {
            "type": "string",
            "format": "uuid",
            "title": "Tenant Id"
          },
          "provider": { "type": "string", "title": "Provider" },
          "client_id": { "type": "string", "title": "Client Id" },
          "client_secret_prefix": {
            "type": "string",
            "title": "Client Secret Prefix"
          },
          "redirect_uri": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Redirect Uri"
          },
          "oidc_discovery_url": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Oidc Discovery Url"
          },
          "is_active": { "type": "boolean", "title": "Is Active" },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "tenant_id",
          "provider",
          "client_id",
          "client_secret_prefix",
          "is_active",
          "created_at",
          "updated_at"
        ],
        "title": "TenantSocialProviderResponse",
        "description": "Response — never includes raw client_secret."
      },
      "TenantSocialProviderToggle": {
        "properties": {
          "is_active": { "type": "boolean", "title": "Is Active" }
        },
        "type": "object",
        "required": ["is_active"],
        "title": "TenantSocialProviderToggle",
        "description": "Body for PATCH /tenants/{tenant_id}/social-providers/{provider}/toggle"
      },
      "TenantSocialProviderUpdate": {
        "properties": {
          "client_id": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Client Id"
          },
          "client_secret": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Client Secret"
          },
          "redirect_uri": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Redirect Uri"
          },
          "oidc_discovery_url": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Oidc Discovery Url"
          },
          "is_active": {
            "anyOf": [{ "type": "boolean" }, { "type": "null" }],
            "title": "Is Active"
          }
        },
        "type": "object",
        "title": "TenantSocialProviderUpdate",
        "description": "Body for PUT /tenants/{tenant_id}/social-providers/{provider}"
      },
      "TenantType": {
        "type": "string",
        "enum": ["PLATFORM", "CUSTOMER"],
        "title": "TenantType"
      },
      "TenantUpdate": {
        "properties": {
          "name": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Name"
          },
          "description": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Description"
          },
          "type": {
            "anyOf": [
              { "$ref": "#/components/schemas/TenantType" },
              { "type": "null" }
            ]
          },
          "owner_id": {
            "type": "string",
            "format": "uuid",
            "title": "Owner Id"
          }
        },
        "type": "object",
        "required": ["owner_id"],
        "title": "TenantUpdate"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": { "anyOf": [{ "type": "string" }, { "type": "integer" }] },
            "type": "array",
            "title": "Location"
          },
          "msg": { "type": "string", "title": "Message" },
          "type": { "type": "string", "title": "Error Type" },
          "input": { "title": "Input" },
          "ctx": { "type": "object", "title": "Context" }
        },
        "type": "object",
        "required": ["loc", "msg", "type"],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": { "HTTPBearer": { "type": "http", "scheme": "bearer" } }
  }
}